feat: convert StatefulButton to TypeScript
- Convert StatefulButton from .jsx to .tsx
- Remove propTypes and defaultProps (React 19 ready)
- Move defaults to function parameter defaults
- Convert test file to TypeScript
- Update src/index.ts to move StatefulButton to typed exports
- Add ESLint override for .tsx files to allow JSX

Addresses #3739 and #3744
